LayoutTests/media/audio-playback-restriction-removed-track-enabled-expected.txt

Test that when RequireUserGestureForAudioRateChange is set, enabling an audio track with a user gesture clears the restriction.


RUN(internals.setMediaElementRestrictions(video, 'RequireUserGestureForAudioRateChange'))
RUN(video.src = findMediaFile('video', 'content/audio-tracks'))
EVENT(canplaythrough)
RUN(video.play())
EVENT(playing)
RUN(audioTrack.enabled = true)
RUN(video.muted = false)
END OF TEST
